refactor(rust): split lib.rs god-module
Decompose the ~2570 LOC lib.rs (audit architecture F1 / maintainability F2)
into focused modules: cli, osc, state, pty, path_security, filesystem,
mcp_settings. Each carries its own #[cfg(test)] tests. lib.rs is now a
162-LOC shell holding run() + the Tauri builder; it re-exports the command
modules so generate_handler! is byte-identical and the command set is
unchanged. Pure structural move; all 81 tests pass.
